+++ psynet/trial/main.py
@@ -508,14 +508,6 @@ class Trial(SQLMixinDallinger, Info):
         if self.trial_maker_id:
             return get_trial_maker(self.trial_maker_id)
 
-    @property
-    def contents(self):
-        return self.answer
-
-    @contents.setter
-    def contents(self, x):
-        self.answer = x
-
     def _allocate_performance_reward(self):
         reward = self.compute_performance_reward(score=self.score)
         assert isinstance(reward, (float, int))